Revert "Add animations + footer"
