+++ package.json
@@ -6,11 +6,8 @@
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/user-event": "^13.5.0",
     "big.js": "^6.2.1",
-    "radium": "^0.26.2",
     "react": "^18.2.0",
-    "react-animations": "^1.0.0",
     "react-dom": "^18.2.0",
-    "react-fade-in": "^2.0.1",
     "react-icons": "^4.4.0",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
