+++ src/App.js
@@ -1,23 +1,19 @@
 import React from 'react';
 import { Routes, Route } from 'react-router-dom';
 
-import Home from './container/homePage/Home';
-import CalculatorPage from './container/calculatorPage/CalculatorPage';
-import Quote from './container/quotePage/Quote';
-import Navbar from './components/navbar/Navbar';
-import Footer from './components/footer/Footer';
+import Home from './container/Home';
+import Calculator from './components/Calculator';
+import Quote from './container/Quote';
+import Navbar from './components/Navbar';
 
 const App = () => (
   <div>
     <Navbar />
-    <main>
-      <Routes>
-        <Route path="/" element={<Home />} />
-        <Route path="/calculator" element={<CalculatorPage />} />
-        <Route path="/quote" element={<Quote />} />
-      </Routes>
-    </main>
-    <Footer />
+    <Routes>
+      <Route path="/" element={<Home />} />
+      <Route path="/calculator" element={<Calculator />} />
+      <Route path="/quote" element={<Quote />} />
+    </Routes>
   </div>
 
 );
