+++ src/components/__tests__/Calculator.test.js
@@ -1,7 +1,7 @@
 import { fireEvent, render, screen } from '@testing-library/react';
 import React from 'react';
 import Renderer from 'react-test-renderer';
-import Calculator from '../calculator/Calculator';
+import Calculator from '../Calculator';
 
 it('renders correctly', () => {
   const tree = Renderer
