+++ src/components/__tests__/Navbar.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Renderer from 'react-test-renderer';
-import Navbar from '../navbar/Navbar';
+import Navbar from '../Navbar';
 
 it('renders correctly', () => {
   const tree = Renderer
