+++ src/components/__tests__/Quote.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Renderer from 'react-test-renderer';
-import Quote from '../../container/quotePage/Quote';
+import Quote from '../../container/Quote';
 
 it('renders correctly', () => {
   const tree = Renderer
