--- src/components/footer/Footer.css
@@ -1,12 +1,0 @@
-footer {
-  position: absolute;
-  bottom: 0;
-  width: 100vw;
-  text-align: center;
-  padding: 2rem 1rem;
-  color: #fff;
-}
-
-footer > a {
-  color: #86c232;
-}
