--- src/components/footer/Footer.js
@@ -1,17 +1,0 @@
-import React from 'react';
-import './Footer.css';
-
-const Footer = () => (
-  <footer>
-    This is a Free/Libre Software under&nbsp;
-    <a
-      href="https://github.com/Mehdi-Rh/Math-magicians"
-      target="_blank"
-      rel="noopener noreferrer"
-    >
-      MIT license
-    </a>
-  </footer>
-);
-
-export default Footer;
