+++ src/container/Quote.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import './Quote.css';
+
+const Quote = () => (
+  <div id="quote">
+    <p>Enjoy Five before Five</p>
+    <p>Your Youth before your Old Age</p>
+    <p>Your Health before your Illness</p>
+    <p>Your Wealth before your Poverty</p>
+    <p>Your Free Time before your Busyness</p>
+    <p>and Your Life before your Death.</p>
+    <p className="author">Prophet Mohammed</p>
+    <p className="author">Peace Upon Him</p>
+  </div>
+);
+
+export default Quote;
