--- src/container/calculatorPage/CalculatorPage.css
@@ -1,32 +1,0 @@
-#calculatorPage {
-  display: flex;
-}
-
-h1 {
-  color: white;
-  margin: 6.5rem 0.5rem 0 4rem;
-}
-
-#calculator {
-  display: inline-grid;
-  margin: 5rem auto auto;
-  font-size: xx-large;
-  justify-content: center;
-}
-
-@media screen and (max-width: 768px) {
-  #calculatorPage {
-    display: flex;
-    flex-direction: column;
-  }
-
-  h1 {
-    color: white;
-    margin: 4rem 0.5rem 0 4rem;
-  }
-
-  #calculator {
-    display: inline-grid;
-    margin: 3rem auto auto;
-  }
-}
