--- src/container/quotePage/Quote.css
@@ -1,56 +1,0 @@
-body {
-  --grText: #61892f;
-  --grBg: #86c232;
-  --grLight: #e5f6df;
-  --black: #222629;
-  --darkGray: #474b4f;
-  --lightGray: #6b6e70;
-}
-
-#quoteContainer {
-  margin: 12vh calc(30vw - 6rem);
-  color: white;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  font-size: x-large;
-}
-
-.quote > * {
-  margin: 0.5rem;
-}
-
-.author {
-  font-size: xx-large;
-  color: var(--grText);
-  display: flex;
-  justify-content: flex-end;
-  margin: 0.5rem;
-}
-
-.mobile {
-  display: none;
-}
-
-@media screen and (max-width: 768px) {
-  #quoteContainer {
-    margin: 10vh;
-    font-size: medium;
-    display: flex;
-  }
-
-  .quote {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-
-  .author {
-    font-size: large;
-    display: flex;
-  }
-
-  .screen {
-    display: none;
-  }
-}
